+++ release/hooks/index.es6.js
@@ -4,3 +4,4 @@ import './useChildrenProvider/index.es6.js';
 import './useResizeListener/index.es6.js';
 import './useClickListener/index.es6.js';
 import './useListener/index.es6.js';
+import './useTouchHide/index.es6.js';