+++ release/hooks/useTouchHide/index.es6.js
@@ -1,0 +1,1 @@
+export { useTouchHide } from './useTouchHide.es6.js';