+++ release/hooks/useTouchHide/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var useTouchHide = require('./useTouchHide.js');
+
+
+
+exports.useTouchHide = useTouchHide.useTouchHide;