+++ release/ui/popups/Drawer/Drawer.d.ts
@@ -1,8 +1,8 @@
+import { TouchHidePlacement } from '../../../hooks';
 import { FlexProps } from '../../layout';
-export declare type DrawerPlacement = 'left' | 'top' | 'right' | 'bottom';
 export interface DrawerProps extends Omit<FlexProps, 'onclose'> {
     size?: number;
-    placement?: DrawerPlacement;
+    placement?: TouchHidePlacement;
     onclose?: (result: string) => void;
 }
-export declare function Drawer({ onclose, size, placement, style, ...props }?: DrawerProps): any;
+export declare function Drawer({ onclose, size, placement, style, ref, ...props }?: DrawerProps): any;