+++ release/ui/popups/Drawer/Drawer.js
@@ -7,10 +7,12 @@ var dom = require('@innet/dom');
 var jsx = require('@innet/jsx');
 var classes = require('html-classes');
 var watchState = require('watch-state');
+require('../../../hooks/index.js');
 require('../../../utils/index.js');
 require('../../icons/index.js');
 require('../../layout/index.js');
 var Drawer$1 = require('./Drawer.scss.js');
+var useTouchHide = require('../../../hooks/useTouchHide/useTouchHide.js');
 var setOverflow = require('../../../utils/setOverflow/setOverflow.js');
 var Flex = require('../../layout/Flex/Flex.js');
 var Icon = require('../../icons/Icon/Icon.js');
@@ -21,18 +23,17 @@ var classes__default = /*#__PURE__*/_interopDefaultLegacy(classes);
 
 const useStyle = dom.style(Drawer$1["default"]);
 let drawersCount = 0;
-const transformPlacements = {
-    left: '-30%, 0',
-    top: '0, -30%',
-    right: '30%, 0',
-    bottom: '0, 30%',
-};
 function Drawer(_a = {}) {
-    var { onclose, size = 388, placement = 'left', style } = _a, props = tslib.__rest(_a, ["onclose", "size", "placement", "style"]);
+    var { onclose, size = 388, placement = 'left', style, ref = new dom.Ref() } = _a, props = tslib.__rest(_a, ["onclose", "size", "placement", "style", "ref"]);
     const children = jsx.useChildren();
     const styles = useStyle();
     const show = dom.useShow();
     const hide = new dom.Ref();
+    const { touched, touchHide, handleTouchStart, handleTouchMove, handleTouchEnd } = useTouchHide.useTouchHide({
+        hide: () => onclose === null || onclose === void 0 ? void 0 : onclose('touch'),
+        placement,
+        element: ref,
+    });
     if (!drawersCount) {
         setOverflow.setOverflow('hidden');
     }
@@ -46,14 +47,16 @@ function Drawer(_a = {}) {
     const overlayHack = {
         _close: () => onclose,
     };
-    return ({type:'delay',props:{ref:hide,hide:300},children:[{type:'div',props:{...overlayHack,style:Object.assign(Object.assign({}, style), { '--ui-drawer-right': placement === 'right' ? '0' : '', '--ui-drawer-bottom': placement === 'bottom' ? '0' : '', '--ui-drawer-width': ['left', 'right'].includes(placement) ? `${size}px` : '', '--ui-drawer-height': ['top', 'bottom'].includes(placement) ? `${size}px` : '', '--ui-drawer-transform': `translate(${transformPlacements[placement]})` }),class:() => {
+    return ({type:'delay',props:{ref:hide,hide:300},children:[{type:'div',props:{...overlayHack,ontouchstart:handleTouchStart,ontouchmove:handleTouchMove,ontouchend:handleTouchEnd,style:Object.assign(Object.assign({}, style), { '--ui-drawer-size': `${size}px`, '--ui-drawer-touch-hide': () => `${touchHide.value}` }),class:() => {
             var _a;
             return classes__default["default"]([
                 styles.root,
                 show.value && styles.show,
+                styles[placement],
+                touched.value && styles.touch,
                 ((_a = hide.value) === null || _a === void 0 ? void 0 : _a.value) && styles.hide,
             ]);
-        }},children:[{type:Flex.Flex,props:{...props,class:() => styles.content},children:[children]},{type:Icon.Icon,props:{onclick:() => onclose === null || onclose === void 0 ? void 0 : onclose('close'),icon:'cross',class:() => styles.close}}]}]});
+        }},children:[{type:Flex.Flex,props:{...props,ref:ref,class:() => styles.content},children:[children]},{type:Icon.Icon,props:{onclick:() => onclose === null || onclose === void 0 ? void 0 : onclose('close'),icon:'cross',class:() => styles.close}}]}]});
 }
 
 exports.Drawer = Drawer;