+++ release/ui/popups/Popout/Popout.es6.js
@@ -7,6 +7,7 @@ import '../../../hooks/index.es6.js';
 import '../../../utils/index.es6.js';
 import '../../layout/index.es6.js';
 import modules_f03e01a5 from './Popout.scss.es6.js';
+import { useTouchHide } from '../../../hooks/useTouchHide/useTouchHide.es6.js';
 import { useEscapeListener } from '../../../hooks/useEscapeListener/useEscapeListener.es6.js';
 import { setOverflow } from '../../../utils/setOverflow/setOverflow.es6.js';
 import { Flex } from '../../layout/Flex/Flex.es6.js';
@@ -22,11 +23,10 @@ function PopoutElement(_a) {
     const hide = useHidden();
     const preshow = useShow();
     const show = useShow(200);
+    const { touched, touchHide, handleTouchStart, handleTouchMove, handleTouchEnd } = useTouchHide({
+        hide: onhide,
+    });
     const styles = useStyle();
-    let touchStart;
-    let touchStartX;
-    const touched = new State(false);
-    const touchHide = new State(0);
     useEscapeListener(onhide);
     const elementData = new State({
         rect: element.value.getBoundingClientRect(),
@@ -52,31 +52,7 @@ function PopoutElement(_a) {
             setOverflow('');
         }
     });
-    return ({type:'div',props:{ref:rootRef,style:Object.assign(Object.assign({}, style), { '--ui-popout-top': () => `${elementData.value.rect.top}px`, '--ui-popout-left': () => `${elementData.value.rect.left}px`, '--ui-popout-width': () => `${elementData.value.rect.width}px`, '--ui-popout-height': () => `${elementData.value.rect.height}px`, '--ui-popout-radius': () => `${elementData.value.styles.borderRadius}`, '--ui-popout-border': () => `${elementData.value.styles.border}`, '--ui-popout-touch-hide': () => `${touchHide.value}px`, '--ui-popout-background': () => `${elementData.value.styles.background}` }),ontouchstart:(e) => {
-            touchStart = e.touches[0].clientY;
-            touchStartX = e.touches[0].clientX;
-            touched.value = true;
-        },ontouchmove:(e) => {
-            if (!touched.value)
-                return;
-            const newTouchHide = e.touches[0].clientY - touchStart;
-            const newTouchStartX = Math.abs(e.touches[0].clientX - touchStartX);
-            if (newTouchStartX > newTouchHide) {
-                touched.value = false;
-                touchHide.value = 0;
-                return;
-            }
-            if (newTouchHide > 100) {
-                touched.value = false;
-                onhide();
-            }
-            else {
-                touchHide.value = e.touches[0].clientY - touchStart;
-            }
-        },ontouchend:() => {
-            touched.value = false;
-            touchHide.value = 0;
-        },class:() => classes([
+    return ({type:'div',props:{ref:rootRef,style:Object.assign(Object.assign({}, style), { '--ui-popout-top': () => `${elementData.value.rect.top}px`, '--ui-popout-left': () => `${elementData.value.rect.left}px`, '--ui-popout-width': () => `${elementData.value.rect.width}px`, '--ui-popout-height': () => `${elementData.value.rect.height}px`, '--ui-popout-radius': () => `${elementData.value.styles.borderRadius}`, '--ui-popout-border': () => `${elementData.value.styles.border}`, '--ui-popout-touch-hide': () => `${touchHide.value}`, '--ui-popout-background': () => `${elementData.value.styles.background}` }),ontouchstart:handleTouchStart,ontouchmove:handleTouchMove,ontouchend:handleTouchEnd,class:() => classes([
             styles.root,
             preshow.value && styles.preshow,
             show.value && styles.show,