+++ release/ui/search/SearchInput/SearchInput.d.ts
@@ -2,4 +2,4 @@ import { InputProps } from '../../interaction';
 export interface SearchInputProps extends Omit<InputProps, 'value'> {
     key: string;
 }
-export declare function SearchInput({ key, oninput, ...props }: SearchInputProps): any;
+export declare function SearchInput({ key, oninput, ...props }: SearchInputProps): import("@innet/jsx").JSXElement<any, import("@innet/jsx").Props, import("@innet/jsx").Children>;